ai-agent-cloud/requirements.txt
# Core dependencies for AI Agent Cloud Infrastructure
openai
python-dotenv

# FastMCP (includes MCP SDK as dependency)
fastmcp

# AWS SDK for Python (Boto3)
boto3
botocore

# Future cloud providers (not yet implemented)
# azure-identity
# azure-mgmt-compute
# azure-mgmt-network
# google-cloud-compute
# google-auth
